Added Oblong Class
Added Oblong Class which is essentially a Rectangle class named
differently to not create ambiguity with
Microsoft.Xna.Framework.Rectangle. It pretty much works in the same way,
however Oblong supports rotation, as with any of the other classes
within the Primitive namespace. Almost everything is implemented, just a
few tweaks and its complete.